consolidate proto utils fns
