+++ kafka-client/src/main/java/dev/responsive/kafka/internal/utils/Utils.java
@@ -12,9 +12,6 @@
 
 package dev.responsive.kafka.internal.utils;
 
-import dev.responsive.kafka.internal.db.rs3.client.LssId;
-import dev.responsive.rs3.Rs3;
-import java.util.UUID;
 import java.util.regex.Pattern;
 import org.apache.kafka.common.utils.Bytes;
 import org.slf4j.Logger;
@@ -128,22 +125,5 @@ public final class Utils {
     } catch (final IndexOutOfBoundsException e) {
       return null;
     }
-  }
-
-  public static UUID uuidFromProto(final Rs3.UUID uuid) {
-    return new UUID(uuid.getHigh(), uuid.getLow());
-  }
-
-  public static Rs3.UUID uuidToProto(final UUID uuid) {
-    return Rs3.UUID.newBuilder()
-        .setHigh(uuid.getMostSignificantBits())
-        .setLow(uuid.getLeastSignificantBits())
-        .build();
-  }
-
-  public static Rs3.LSSId lssIdProto(final LssId lssId) {
-    return Rs3.LSSId.newBuilder()
-        .setId(lssId.id())
-        .build();
   }
 }
